claim amount-based fraud trigger

For each procedure there is a specific amount that is generally agreed between the insurer and the medical service provider. This is especially the case for cashless claim processing. If the claim amount is higher than this agreed procedure specific amount it will raise a flag.

Usage

claim_amount_trigger_map(
  claims_file,
  claim_paid_field,
  triggers_file,
  procedure_code,
  procedure_amount_field
)

Arguments

claims_file

data set containing claim-level information

claim_paid_field

column in the claims_file that contains the claim amount paid by the insurer

triggers_file

file containing the agreed specific amount for each procedure

procedure_code

column in the triggers_file that contains the code/identifier for the medical procedure

procedure_amount_field

column in the triggers_file that contains the agreed specific amount for each procedure

Value

input claims_file with claim_amount_ratio and claim_amount_flag columns

Examples

library(dplyr)

data(claims_data_sample)
data(trigger_data)

claim_amount_trigger_map(
claims_file = claims_data_sample,
claim_paid_field = "approved_allowed_amount",
triggers_file = trigger_data,
procedure_code = "primary_procedure_code",
procedure_amount_field = "package_amount_trigger") %>%
select(c("approved_allowed_amount", "claim_amount_ratio", "claim_amount_flag"))
#> # A tibble: 1,000 × 3
#>    approved_allowed_amount claim_amount_ratio claim_amount_flag
#>                      <dbl>              <dbl>             <dbl>
#>  1                   15300              1.99                  1
#>  2                   13500              1.8                   1
#>  3                   16200              0.814                 1
#>  4                   10800              1.17                  1
#>  5                    5400              0.651                 1
#>  6                   23400              0.992                 1
#>  7                   10800              0.543                 1
#>  8                   16200              1.95                  1
#>  9                   10800              1.44                  1
#> 10                    1800              0.196                 1
#> # ℹ 990 more rows