Gender-based fraud trigger

Some procedures are gender specific. This function creates a red flag if a procedure is performed for a claimant for a wrong gender. For example gynacologic procedure for a male claimant will raise a redflag. The gender should be specified consistently in the claims_file and the triggers_file for this function to work.

Usage

gender_trigger_map(
  claims_file,
  gender_column_name,
  triggers_file,
  procedure_code,
  gender_trigger
)

Arguments

claims_file

data set containing claim-level information

gender_column_name

column in the claims_file that contains the gender of the policy holder

triggers_file

file containing the gender specification for each procedure code

procedure_code

column in the triggers_file that contains the code/identifier for the medical procedure

gender_trigger

column in the triggers_file that contains the specification of gender for each medical procedure

Value

input claims_file with gender_flag column

Examples

library(dplyr)

data(claims_data_sample)
data(trigger_data)

gender_trigger_map(
claims_file = claims_data_sample,
gender_column_name = "gender_code",
triggers_file = trigger_data,
procedure_code = "primary_procedure_code",
gender_trigger = "gender_trigger") %>%
group_by(gender_code)%>%
summarise(gender_flag = mean(gender_flag))
#> # A tibble: 2 × 2
#>   gender_code gender_flag
#>   <chr>             <dbl>
#> 1 Female           0     
#> 2 Male             0.0494